Renames db files
